# [aggregator] Checking if metadata is set to default should not cause …